app/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

